+++ apps/designer/app/routes/login.tsx
@@ -1,6 +1,6 @@
 import type { LoaderArgs, MetaFunction } from "@remix-run/node";
 import { Root } from "~/shared/remix";
-import env from "~/env/env.public";
+import env from "~/env/env.public.server";
 import { getThemeData } from "~/shared/theme";
 
 export const loader = async ({ request }: LoaderArgs) => {
